pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gioe-libs"
version = "0.1.0"
description = "Shared Python libraries for common service patterns (logging, alerting, observability, cron)"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
observability = [
    "sentry-sdk[opentelemetry]>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-exporter-prometheus>=0.41b0",
]

[tool.setuptools]
packages = [
    "gioe_libs",
    "gioe_libs.alerting",
    "gioe_libs.cron_runner",
    "gioe_libs.domain_types",
    "gioe_libs.aiq_logging",
    "gioe_libs.observability",
]

[tool.setuptools.package-dir]
"gioe_libs" = "."
